fix: sleep details, HRV fallback, workout calories float
- Call /sleep endpoint for detailed durations (rem/deep/light/awake etc)
  aggregated by day; /daily_sleep only returns score
- Fall back to sleep-session average_hrv when /daily_hrv returns 404
- Fix workout calories type *int → *float64 (Oura returns float values)
- Migration 004 alters workouts.calories from INT to FLOAT

+++ backend/models/workout.go
@@ -12,7 +12,7 @@ type Workout struct {
 	StartDatetime time.Time  `db:"start_datetime" json:"start_datetime"`
 	EndDatetime   *time.Time `db:"end_datetime" json:"end_datetime"`
 	Activity      *string    `db:"activity" json:"activity"`
-	Calories      *int       `db:"calories" json:"calories"`
+	Calories      *float64   `db:"calories" json:"calories"`
 	Distance      *float64   `db:"distance" json:"distance"`
 }
